src/blog/blog.service.ts
import { Injectable } from '@nestjs/common';
import { InjectRepository } from '@nestjs/typeorm';
import { UserEntity } from '../user/user.entity';
import { BlogRepository } from './blog.repository';
import { BlogFilter } from './types/blog.filter';
import { BlogInputType } from './types/blog.input';

@Injectable()
export class BlogService {
  constructor(
    @InjectRepository(BlogRepository)
    private blogRepository: BlogRepository,
  ) {}

  async createBlog(user: UserEntity, input: BlogInputType) {
    return this.blogRepository.createBlog(user, input);
  }

  async createOrUpdateBlog(input: BlogInputType, user: UserEntity) {
    return this.blogRepository.createOrUpdateBlog(input, user);
  }

  async getBlogById(id: string) {
    return this.blogRepository.getBlogById(id);
  }
  async allBlogs(input: BlogFilter) {
    return this.blogRepository.allBlogs(input);
  }

  async blogs(user: UserEntity) {
    return this.blogRepository.getBlogs(user);
  }

  async updateBlog(input: BlogInputType) {
    return this.blogRepository.updateBlog(input);
  }

  async deleteBlog(id: string) {
    return this.blogRepository.deleteBlog(id);
  }
}
